+++ egui_demo_lib/src/wrap_app.rs
@@ -54,10 +54,6 @@ impl epi::App for WrapApp {
         self.backend_panel.max_size_points_active
     }
 
-    // Let's show off that we support transparent windows (on native):
-    fn transparent(&self) -> bool {
-        true
-    }
     fn clear_color(&self) -> egui::Rgba {
         egui::Rgba::TRANSPARENT // we set a `CentralPanel` fill color in `demo_windows.rs`
     }