+++ eframe/CHANGELOG.md
@@ -3,6 +3,7 @@ All notable changes to the `eframe` crate.
 
 
 ## Unreleased
+* Moved options out of `trait App` into new `NativeOptions`.
 
 
 ## 0.11.0 - 2021-04-05